pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bazarrbulksync"
version = "0.2.0"
description = "Bulk subtitle synchronization for Bazarr"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Brian Wei Hao Ma", email = "brianmaytc@gmail.com"}]
classifiers=[
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Topic :: Multimedia :: Video",
    "Intended Audience :: System Administrators",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.7",
    "typer>=0.12",
]
keywords = [
    "bazarr",
    "subtitles",
    "subtitle-sync",
    "subtitle-synchronization",
    "media",
    "video",
    "movies",
    "tv",
    "cli",
    "command-line",
    "automation",
    "bulk",
    "sync",
    "media-server",
    "plex",
    "emby",
    "jellyfin",
    "sonarr",
    "radarr",
    "streaming",
    "bulk-operation",
    "docker",
    "python",
    "synchronization",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
bazarrbulksync = "bazarrbulksync.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/bazarrbulksync"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
